Alignment with Java suite implementation
