+++ lib/js/src/manager/lifecycle/_LifecycleManager.js
@@ -91,6 +91,8 @@ class _LifecycleManager {
         this._systemCapabilityManager = new SystemCapabilityManager(this);
         this._encryptionLifecycleManager = null;
         this._registerAppInterfaceResponse = null;
+
+        this._didCheckSystemInfo = false;
     }
 
     /**
@@ -100,12 +102,21 @@ class _LifecycleManager {
      */
     _createSessionListener () {
         const sessionListener = new _SdlSessionListener();
-        sessionListener.setOnProtocolSessionStarted((serviceType, sessionID, version, correlationID, hashID, isEncrypted) => {
+        sessionListener.setOnProtocolSessionStarted((serviceType, sessionID, version, correlationID, hashID, isEncrypted, systemInfo) => {
             // Session has been started
             if (this._minimumProtocolVersion !== null && this._minimumProtocolVersion.isNewerThan(this.getProtocolVersion()) === 1) {
                 console.warn(`Disconnecting from head unit, the configured minimum protocol version ${this._minimumProtocolVersion} is greater than the supported protocol version ${this.getProtocolVersion()}`);
                 this._sdlSession.endService(serviceType, this._sdlSession.getSessionId());
                 return this._cleanProxy();
+            }
+
+            if (systemInfo && !this._didCheckSystemInfo) {
+                this._didCheckSystemInfo = true;
+                if (!this._onSystemInfoReceived(systemInfo)) {
+                    console.warn('Disconnecting from head unit, the system info was not accepted.');
+                    this._sdlSession.endService(serviceType, this._sdlSession.getSessionId());
+                    return this._cleanProxy();
+                }
             }
 
             if (serviceType === _ServiceType.RPC) {
@@ -509,10 +520,8 @@ class _LifecycleManager {
             this._cleanProxy();
         }
 
-        // to avoid double firing the onSystemInfoReceived event
-        // try to define systemInfo and fire onSystemInfoReceived here
-        // only if that was not checked in SdlSession before
-        if (!this._sdlSession.didCheckSystemInfo()) {
+        if (!this._didCheckSystemInfo) {
+            this._didCheckSystemInfo = true;
             const systemInfo = new SystemInfo(registerAppInterfaceResponse.getVehicleType(), registerAppInterfaceResponse.getSystemSoftwareVersion()) ;
             if (!this.onSystemInfoReceived(systemInfo)) {
                 console.warn('Disconnecting from head unit, the system info was not accepted.');
