+++ lib/js/src/protocol/_SdlProtocolBase.js
@@ -45,9 +45,6 @@ import { RpcCreator } from '../rpc/RpcCreator.js';
 import { ImageResolution } from '../rpc/structs/ImageResolution.js';
 import { VideoStreamingFormat } from '../rpc/structs/VideoStreamingFormat.js';
 
-import { VehicleType } from '../rpc/structs/VehicleType.js';
-import { SystemInfo } from '../util/SystemInfo.js';
-
 /**
  * Base implementation of sdl protocol.
  * Should be able to handle basic control frames and be able to
@@ -272,37 +269,6 @@ class _SdlProtocolBase {
     }
 
     /**
-     * Extracts the SystemInfo out of a packet
-     * @param {_SdlPacket} sdlPacket - should be a StartServiceACK for the RPC service.
-     * @returns {SystemInfo|null} - an instance of SystemInfo if the information is available, null otherwise.
-     */
-    _extractSystemInfo (sdlPacket) {
-        const make = sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.MAKE);
-        const model = sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.MODEL);
-        const modelYear = sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.MODEL_YEAR);
-        const trim = sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.TRIM);
-
-        const systemHardwareVersion = sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.SYSTEM_HARDWARE_VERSION);
-        const systemSoftwareVersion = sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.SYSTEM_SOFTWARE_VERSION);
-
-        // if no any VehicleType tags in the packet just return null
-        let vehicleType = null;
-        if (make || model || modelYear || trim) {
-            vehicleType = new VehicleType({
-                make,
-                model,
-                modelYear,
-                trim,
-            });
-        }
-        if (!vehicleType && !systemSoftwareVersion && !systemHardwareVersion) {
-            return null;
-        }
-
-        return new SystemInfo(vehicleType, systemSoftwareVersion, systemHardwareVersion);
-    }
-
-    /**
      * Takes an rpc message and sends a single or multi frame packets.
      * @param {RpcMessage} rpcMessage - Converts an RpcMessage into an _SdlPacket and sends it.
      */
@@ -470,22 +436,6 @@ class _SdlProtocolBase {
                         sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.AUTH_TOKEN)
                     );
                 }
-
-                if ((this._protocolVersion.isNewerThan(new Version(5, 4, 0)) >= 0)) {
-                    // check if systemInfo data exists then fire onSystemInfoReceived protocol listener event
-                    const systemInfo = this._extractSystemInfo(sdlPacket);
-                    if (
-                        systemInfo
-                        && this._sdlProtocolListener !== null
-                        && typeof this._sdlProtocolListener.onSystemInfoReceived === 'function'
-                    ) {
-                        if (!this._sdlProtocolListener.onSystemInfoReceived(systemInfo)) {
-                            console.warn('Disconnecting from head unit, the system info was not accepted.');
-                            this.endSession();
-                            return;
-                        }
-                    }
-                }
             } else if (serviceType === _ServiceType.VIDEO) {
                 if (this._sdlProtocolListener !== null) {
                     const acceptedResolution = new ImageResolution();
@@ -512,7 +462,7 @@ class _SdlProtocolBase {
         }
 
         this._sdlProtocolListener.onProtocolSessionStarted(serviceType,
-            sdlPacket.getSessionID(), this._protocolVersion.getMajor(), '', this._hashID, sdlPacket.getEncryption());
+            sdlPacket.getSessionID(), this._protocolVersion.getMajor(), '', this._hashID, sdlPacket.getEncryption(), sdlPacket);
     }
 
     /**
