+++ lib/js/src/protocol/_SdlProtocolListener.js
@@ -57,7 +57,6 @@ class _SdlProtocolListener {
         this._onProtocolSessionEnded = null;
         this._onProtocolSessionEndedNACKed = null;
         this._onAuthTokenReceived = null;
-        this._onSystemInfoReceived = null;
         this._getSessionId = null;
         this._onTransportConnected = null;
     }
@@ -113,16 +112,17 @@ class _SdlProtocolListener {
 
     /**
      * Safely attempts to invoke the event listener.
-     * @param {ServiceType} serviceType - A ServiceType enum value.
+     * @param {_ServiceType} serviceType - A ServiceType enum value.
      * @param {Number} sessionId - A session ID.
      * @param {Number} version - A numeric version.
      * @param {String} correlationId - A correlation ID.
      * @param {Number} hashId - A hash ID.
      * @param {Boolean} isEncrypted - Whether or not it is encrypted.
+     * @param {_SdlPacket} sdlPacket - An _SdlPacket.
      */
-    onProtocolSessionStarted (serviceType, sessionId, version, correlationId, hashId, isEncrypted) {
+    onProtocolSessionStarted (serviceType, sessionId, version, correlationId, hashId, isEncrypted, sdlPacket) {
         if (typeof this._onProtocolSessionStarted === 'function') {
-            this._onProtocolSessionStarted(serviceType, sessionId, version, correlationId, hashId, isEncrypted);
+            this._onProtocolSessionStarted(serviceType, sessionId, version, correlationId, hashId, isEncrypted, sdlPacket);
         }
     }
 
@@ -138,7 +138,7 @@ class _SdlProtocolListener {
 
     /**
      * Safely attempts to invoke the event listener.
-     * @param {ServiceType} serviceType - A ServiceType enum value.
+     * @param {_ServiceType} serviceType - A ServiceType enum value.
      * @param {Number} sessionId - A Session ID.
      * @param {String} correlationId - A correlation ID.
      */
@@ -160,7 +160,7 @@ class _SdlProtocolListener {
 
     /**
      * Safely attempts to invoke the event listener.
-     * @param {ServiceType} serviceType - A ServiceType enum value.
+     * @param {_ServiceType} serviceType - A ServiceType enum value.
      * @param {Number} sessionId - A Session ID.
      * @param {String} correlationId - A correlation ID.
      */
@@ -188,28 +188,6 @@ class _SdlProtocolListener {
         if (authToken !== null) {
             this._onAuthTokenReceived(authToken);
         }
-    }
-
-    /**
-     * Set the OnSystemInfoReceived function callback.
-     * @param {function} listener - A function to be invoked when the event occurs.
-     * @returns {_SdlProtocolListener} - A reference to this instance to support method chaining.
-     */
-    setOnSystemInfoReceived (listener) {
-        this._onSystemInfoReceived = listener;
-        return this;
-    }
-
-    /**
-     * Safely attempts to invoke the onSystemInfoReceived event.
-     * @param {SystemInfo} systemInfo - the system info of vehicle that this session is currently active on.
-     * @returns {Boolean} Return true if this session should continue, false if the session should end
-     */
-    onSystemInfoReceived (systemInfo) {
-        if (typeof this._onSystemInfoReceived === 'function') {
-            return !!this._onSystemInfoReceived(systemInfo);
-        }
-        return true;
     }
 
     /**
