+++ lib/js/src/session/_SdlSession.js
@@ -35,6 +35,10 @@ import { _SdlProtocol } from '../protocol/_SdlProtocol.js';
 import { _ServiceType } from '../protocol/enums/_ServiceType.js';
 import { _ServiceListenerMap } from './_ServiceListenerMap.js';
 import { _VideoStreamingParameters } from '../streaming/video/_VideoStreamingParameters.js';
+import { _ControlFrameTags } from '../protocol/enums/_ControlFrameTags.js';
+import { VehicleType } from '../rpc/structs/VehicleType.js';
+import { SystemInfo } from '../util/SystemInfo.js';
+import { Version } from '../util/Version.js';
 
 
 /**
@@ -61,7 +65,6 @@ class _SdlSession {
         this._sdlProtocolListener = this._setupSdlProtocolListener();
 
         this._sdlProtocol = new _SdlProtocol(baseTransportConfig, this._sdlProtocolListener);
-        this._didCheckSystemInfo = false;
     }
 
     /**
@@ -78,12 +81,41 @@ class _SdlSession {
         sdlProtocolListener.setOnRpcMessageReceived(this.onRpcMessageReceived.bind(this));
         sdlProtocolListener.setOnTransportConnected(this.onTransportConnected.bind(this));
         sdlProtocolListener.setOnAuthTokenReceived(this.onAuthTokenReceived.bind(this));
-        sdlProtocolListener.setOnSystemInfoReceived(this.onSystemInfoReceived.bind(this));
 
         sdlProtocolListener.setGetDesiredVideoParams(this.getDesiredVideoParams.bind(this));
         sdlProtocolListener.setSetAcceptedVideoParams(this.setAcceptedVideoParams.bind(this));
 
         return sdlProtocolListener;
+    }
+
+    /**
+     * Extracts the SystemInfo out of a packet
+     * @param {_SdlPacket} sdlPacket - should be a StartServiceACK for the RPC service.
+     * @returns {SystemInfo|null} - an instance of SystemInfo if the information is available, null otherwise.
+     */
+    _extractSystemInfo (sdlPacket) {
+        const make = sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.MAKE);
+        const model = sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.MODEL);
+        const modelYear = sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.MODEL_YEAR);
+        const trim = sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.TRIM);
+
+        const systemHardwareVersion = sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.SYSTEM_HARDWARE_VERSION);
+        const systemSoftwareVersion = sdlPacket.getTag(_ControlFrameTags.RPC.StartServiceACK.SYSTEM_SOFTWARE_VERSION);
+
+        let vehicleType = null;
+        if (make || model || modelYear || trim) {
+            vehicleType = new VehicleType({
+                make,
+                model,
+                modelYear,
+                trim,
+            });
+        }
+        if (!vehicleType && !systemSoftwareVersion && !systemHardwareVersion) {
+            return null;
+        }
+
+        return new SystemInfo(vehicleType, systemSoftwareVersion, systemHardwareVersion);
     }
 
     /**
@@ -120,15 +152,21 @@ class _SdlSession {
      * @param {String} correlationId - A correlationID.
      * @param {Number} hashId - A hash ID.
      * @param {Boolean} isEncrypted - Whether or not it is encrypted.
+     * @param {_SdlPacket} sdlPacket - An _SdlPacket.
      */
-    onProtocolSessionStarted (serviceType, sessionId, version, correlationId, hashId, isEncrypted) {
+    onProtocolSessionStarted (serviceType, sessionId, version, correlationId, hashId, isEncrypted, sdlPacket) {
         this._sessionId = sessionId;
 
         if (serviceType === _ServiceType.RPC) {
             this._sessionHashId = hashId;
         }
 
-        this._sdlSessionListener.onProtocolSessionStarted(serviceType, sessionId, version, correlationId, hashId, isEncrypted);
+        let systemInfo = null;
+        if (version && (this.getProtocolVersion().isNewerThan(new Version(5, 4, 0)) >= 0)) {
+            systemInfo = this._extractSystemInfo(sdlPacket);
+        }
+
+        this._sdlSessionListener.onProtocolSessionStarted(serviceType, sessionId, version, correlationId, hashId, isEncrypted, systemInfo);
         this._serviceListeners.sendEventServiceStarted(this, serviceType, isEncrypted);
     }
 
@@ -170,25 +208,6 @@ class _SdlSession {
         this._sdlSessionListener.onAuthTokenReceived(authToken, this._sessionId);
     }
 
-    /**
-     * A way to determine if SystemInfo already checked
-     * @returns {Boolean} Return true if received
-     */
-    didCheckSystemInfo () {
-        return this._didCheckSystemInfo;
-    }
-
-    /**
-     * A way to determine if this SDL session should continue to be active while
-     * connected to the determined vehicle type.
-     * @param {SystemInfo} systemInfo - the type of vehicle that this session is currently active on.
-     * @returns {Boolean} Return true if this session should continue, false if the session should end
-     */
-    onSystemInfoReceived (systemInfo) {
-        // set the flag as this event fires only once if VehicleType was received
-        this._didCheckSystemInfo = true;
-        return this._sdlSessionListener.onSystemInfoReceived(systemInfo);
-    }
 
     /** **********************************************************************************************************************************************************************
      * END:    _SdlProtocolListener implemented methods
