+++ lib/js/src/session/_SdlSessionListener.js
@@ -47,7 +47,6 @@ class _SdlSessionListener {
         this._onRpcMessageReceived = null;
         this._onTransportConnected = null;
         this._onAuthTokenReceived = null;
-        this._onSystemInfoReceived = null;
     }
 
     /**
@@ -102,22 +101,23 @@ class _SdlSessionListener {
 
     /**
      * Safely attempts to invoke the onProtocolSessionStarted event.
-     * @param {ServiceType} serviceType - The ServiceType.
+     * @param {_ServiceType} serviceType - The ServiceType.
      * @param {Number} sessionID - represents a byte
      * @param {Number} version - represents a byte
      * @param {String} correlationID - The correlationID.
      * @param {Number} hashID - The hash ID.
      * @param {Boolean} isEncrypted - Whether or not it is encrypted.
+     * @param {SystemInfo} systemInfo - the system info of vehicle that this session is currently active on.
      */
-    onProtocolSessionStarted (serviceType, sessionID, version, correlationID, hashID, isEncrypted) {
+    onProtocolSessionStarted (serviceType, sessionID, version, correlationID, hashID, isEncrypted, systemInfo) {
         if (typeof this._onProtocolSessionStarted === 'function') {
-            this._onProtocolSessionStarted(serviceType, sessionID, version, correlationID, hashID, isEncrypted);
+            this._onProtocolSessionStarted(serviceType, sessionID, version, correlationID, hashID, isEncrypted, systemInfo);
         }
     }
 
     /**
      * Safely attempts to invoke the onProtocolSessionEnded event.
-     * @param {ServiceType} serviceType - The ServiceType.
+     * @param {_ServiceType} serviceType - The ServiceType.
      * @param {Number} sessionID - represents a byte
      * @param {String} correlationID - The correlationID.
      */
@@ -129,7 +129,7 @@ class _SdlSessionListener {
 
     /**
      * Safely attempts to invoke the onProtocolSessionEndedNACKed event.
-     * @param {ServiceType} serviceType - The ServiceType.
+     * @param {_ServiceType} serviceType - The ServiceType.
      * @param {Number} sessionID - represents a byte
      * @param {String} correlationID - The correlationID.
      */
@@ -158,28 +158,6 @@ class _SdlSessionListener {
         if (authToken !== null && typeof this._onAuthTokenReceived === 'function') {
             this._onAuthTokenReceived(authToken, sessionId);
         }
-    }
-
-    /**
-     * Set the onSystemInfoReceived function.
-     * @param {function} listener - A function to be invoked when the event occurs.
-     * @returns {_SdlSessionListener} - A reference to this instance to allow method chaining.
-     */
-    setOnSystemInfoReceived (listener) {
-        this._onSystemInfoReceived = listener;
-        return this;
-    }
-
-    /**
-     * Safely attempts to invoke the onSystemInfoReceived event.
-     * @param {SystemInfo} systemInfo - the system info of vehicle that this session is currently active on.
-     * @returns {Boolean} Return true if this session should continue, false if the session should end
-     */
-    onSystemInfoReceived (systemInfo) {
-        if (typeof this._onSystemInfoReceived === 'function') {
-            return !!this._onSystemInfoReceived(systemInfo);
-        }
-        return true;
     }
 
     /**
